Updating keep-alive/fee transaction
The changes in the transaction generator (which is only for development purposes) in worker.go just mask the issue. If you change the 0chain.yaml configuration so: max_transactions: 1, min_block_size: 100, and max_block_size: 200, then the chain will stall in the first round. The same thing would happen if there are no transactions. To address this we need to change line 191 in the protocol_block.go from OR to AND. Before, we used OR because we wanted to guarantee a minimum block size. However, if there aren't enough transactions to meet the minimum or there are zero transactions then the chain will stall and the miners will not receive any rewards for block generation. If we change it to AND then the miners will wait for a certain time then just create a block with whatever they have. Even if there are zero transactions the miner will create a fee transaction after line 191 so they can get paid then every block will have at least one transaction in it. What was the configuration when you saw it unstable? I've tested it with 0 transactions and 70 transactions for 10000 rounds each and there were no issues.