Update protocol_view_change.go
Allows miners to continue with dkg protocol if there is an acceptable error.

Example: Force one miner to always reply with an error during the share or sign phase. That will delay the view change indefinitely.

By allowing the miners to continue without receiving a sign of the share from every miner the view change can continue with an acceptable error.