Fix issue of GetLatestFinalizedMagicBlockFromSharders
The issue exist in sorting the returned latest magic blocks from sharders. The blocks may all have the same magic block StartingRound, however, the block itself may on different round. If the sharders have not been initialized or catched up with the network, they may  not have the round info. If the block with 0 round number was picked as the latest finalized magic block from sharders, then issues would occur when verify chain history. And may cause BC stuck.

The fix is rewrite the sorting function, which should take the situation of when magic block starting round are all the same into consideration. And if so, we need to compare the block's Round number then. As long as we have at least one sharder with the correct latest finalized magic block, we should be fine. Otherwise, the BC must be stuck to avoid incorrect finalized magic block.