Restart status monitor if the current magic block changed
Even when magic block starting round is the same, the magic block may
have changed, and if we don't restart the status monitor, it would inactive miners
or sharders in the old pools, hence the node active status would not be updated
in current magic block. This will lead further issue for calculating the TKN value
when create new magic block.