Fix DKG process stuck
In move function for contribute, the mpk keys is checked to meet the condition that len(mpks) >= dkgMinersList.K. In the following phase, sos.SignOrShare will be computed base on the mpks, one thing that need to note is the miner's own share will not be included, hence if the current mpks is 7, the sos will be 6. In the previous code, the condition for passing the phase is sos num >= dkgMinersList.N - 2, which is incorrect. As the max number of sos it can be is len(mpks) - 1, however, miners could move to current phase if the len(mpks) >= dkgMinersList.K. Therefore, if the len(mpks) == K, then len(sos) will be K - 1, as long as the N - 2 > K - 1, i.e  N - 1 > K, the DKG process will fail.