Fix MPT path issue
If two leaf nodes on different paths have the same value, and if they have the same ending path byte, they might generate the same node key. If we delete one node on one path, another lead node will be deleted as well.

In this commit, we will add a `prefix` field to the leaf node struct so that the leaf nodes would generate different keys(as they prefix paths are differnt) and hence to avoid the node key conflict on db. We don't add prefix path to other node types such as `Extension` and `Full` node. That's because they have child node key fields, which helps them to avoid the keys conflict.