Cache encoded and/or compressed data when sending requests
This is for improving the performance for broadcasting data. Previously, when we do broadcasting, the `entity` will be encoded and compressed to bytes slice N times for N miners, which is a waste of resources. This commit fixes the problem, and entity will only be encoded and compressed once now.