Fix a concurrent issue on updating RRS of a round
The variables of `seed` and `hasRandomSeed` of Round struct are two atomic values, but the `minerRanks` is a slice protected by a mutex. Previously, when we set the seed to a Round, we update the `seed` and `hasRandomSeed` first, then update the minerRanks base on the seed. The problem is after we update the `seed` and `hasRandomSeed`, but before updating the minerRanks, if there are other goroutines that are checking the Round, it will find that the Round has seed ready, then it will try to access the minerRanks slice, and get panic because it has not been updated yet.