Fix chain split on current round that caused chain stuck
In some situation, the chain could get stuck because the miners has different RRS on the current round. That's because they are using different notarized block of the round. This can cause the chain stuck, if the (threshold - 1) miners have RRS-1, while the remaining miners are have RRS-2, then the next RRS will never be able to reach the threshold.

The good news is even those miners are having different RRS of current round, but they all have the same heaviest notarized block of that round, so we can fix this by resetting the round's RRS to the heaveist notarized block's RRS when round timeout occurs.