Fix a bug in restartRound that could cause chain stuck
Chain got stuck because there are two pars of miners that are redoing VRF share on different round and none of the part could meet the VRF share threshold. The way to fix is sync blocks up in restartRound until no notarized blocks could be acquired. In this way, all miners could have the same page of the latest notarized block, then redo VRF share on the for the same round.