Fix incorrect ComputeFinalizedBlock()
Change back to the old version, and add the previous blocks checking to see if they have notarized blocks instead of returning error directly when it finds the current round does not have notarized block. In this way, each time when we call the finalizeRound() function, we will have the chance to finalize block that is near the latest finalize block.

The problem of the previous version is it returns the heaviest notarized block of a round directly, which may not able to link to the latest finalized block. The old version will get all the notarized blocks, and check if they have the same previous block, if so, the previous block will be returned. Otherwise, it will continue to check the previous blocks till it finds the common block or reached nil.