Fix block state issue in ApplyBlockStateChange
When applying state changes that are acquired from remote, if a block's previous block does exist and its state is computed, then we should link the current block's state db to the previous one. Otherwise,the block will not be able to find MPT nodes in previous blocks, even the miner does have.