Update finalize round process
In finalize round process, we will calculate the new LFB round first, then start to collect all the missing blocks from the new LFB back to the previous LFB. The problem is if the previous LFB is more than 100 rounds older than the new LFB, the new LFB will never be able to reach it then, as all miners will delete blocks that are 100 rounds before. Even if the round could reach to it, it is too much of work for CPU to fetch and verify all of those missing blocks.

In this commit, we will set the max number of rounds we will check back for calculating the new LFB. Also, we will set the max backwards depth from the new LFB to avoid fetching blocks too much.