Setup block verify worker
Previously, a miner will create a new goroutine for every `verify block` message it receives. This could make the miner get stuck without the ability to get recovered. In this commit, we will queue `verify block` message in a buffered channel, and have numbers of workers to get blocks from the queue and process them. When the buffer is full, the new message will be dropped if it could not be pushed to the channel in specific time range.