Fix state save slow [development env only]
Finalizing blocks process on sharders got stuck due to the slow MPT state changes saving. This happened when I changed the miners number from 12 to 24 on hero server. The state saving process is slow and sometimes takes more than 10 seconds to finish. The reason of why it is so slow is it calls the `Flush()` function everytime after it write the state changes to rocksdb. This may not be a problem if the miner runs on a server that do not have frequently disk I/O requests. But apparently, it is a problem when I double the miners number from 12 to 24 miners on a single server.

For rocksdb, any changes write to the db will not be committed to disk immediately. Instead, they will be saved in a memory buffer, and flushed to disk when the buffer is full, or we call the `Flush()` function.

In this commit, `Flush()` on every change is turned off, this could be a problem that may cause state changes data losing if sharders and miners are closed all at the same time. Usually this would not happen, but we'd better enable the `Flush()` on every changes in production env.

TODO: make the calling of `Flush()` on every changes configurable, and enable it by default.