Fix goroutines leak issue on ToN2NReceiveEntityHandler()
After increasing the max block size and max generation transactions, some miners got goroutine leaks issue and can not recover from it. The main issue is the `SetPublicKey()` was called in `ComputeClientID()` function, while it is not necessary at all for computing a client ID. We can compute ID from the public key string directly.

The `ComputeClientID()` function will be called when decoding a transaction, so the more transactions a block has, the more times the function will be called when decoding a block in the `ToN2NReceiveEntityHandler()`, therefore, the more times the `SetPublicKey()` function will be called previously. After fixing this, miners work much stable now.