Add the capacity for canceling the ValidateTransactions()
Previously, when a block is notarized, we will call CancelBlockVerification to stop the block verification process. It can cancel the goroutine for accepting new block verify requests, but the ValidateTransactions() in VerifyBlock() is not cancelable. This could be a waste of CPU resources, as validating the transactions in the block is useless.

Also, since we do not verify blocks just in time when they are proposed, it is possible blocks on different rounds are proposed and the miner may validate transactions in different blocks concurrently. This leads to the high usage of CPU when the goroutines number for ValidateTransactions increased. When the current round is moved on, all of those ValidateTransactions() are useless and we should be able to cancel them immediately.