Init constants, volumes, cold storages and register read and write function
Sharder can provide different combination of storages and accordingly read and write functions are registered and will be called when block read and write operation is done.

This change calls initialization function that initializes storages and registers strategy which will be called when block write is called.

Blocks metadata is stored is boltdb which can be recovered when disk containing boltdb fails. Initializaiton should be run in "recover" move to recover metadata.