Fix http response data reading closed early due to timeout
Previously, we use timeout context for the http request, there is a problem about this when the sender is slow on reading the response data. If we can not read out all the data in the response in before timeout, we will then get error of `context deadline exceeded`. What we should do is use context.WithCancel(), and have another timer to call the cancel() function if we don't get any response in time, and stop the timer once we have get response successfully, and then start to read out data from the response.