Simplify pull data process
Seems the complicate process for pulling data from remote is unnecessary, a miner should pull the data from the node who sent the request directly instead of push the node into an array, and start to try to pull data from node one by one from the array. It's useless and if the first node is not the one who sent the request, it just does not have the data the miner needs.