Optimize signature signing in sending requests
Suppose we have 100 miners, when we send a message, we need to sign the same message 100 times with different timestamps to 100 nodes. This could be optimized, as the message is the same, the only differences is the timestamp. We can not use the same timestamp for all the signatures, as the sending process takes time, it could take more than 4 seconds to send data to all the 100 miners, the timestamp will be expired when the last node received the message. So the way to optimize is we pre-generate 5 signatures that is signed with the same message, but with different timestamps. So the first signature is sign(msg,t_now), the second will be sign(msg, t_now + 1 second), and so on. When we set the signature to the http request header, we will use the signature that has timestamp 1 second within the current time. If all the pre-generated signatures' timestamp have expired, we will generate a new one then.