Fix error on pulling 'block/notarization' data
Continually saw 'Requested data is not found' error on pulling 'block/notarization' data. Force set the data type from `push` to `pull` mode fixed this issue. However,
there may still have issues on pulling mode for getting data from remote. The condition for triggering the `pullEntityHandler` might be buggy, it calls the `pullEntityHandler` to pull data from remote when it found no data in the response. So if the data is not big enough to switch to `pull` mode, then no data will be cached for later pulling. And if the receiver does not found data in the response, then it will try to pull the data from the sender, apparently, the sender does not have the data in the cache, then we will see the error above.