Force stop and do fast moving LFB when it is behind too much
When a miner spends too much time on finalizing a block, it takes time and resources to sync and finalize all blocks between the finalizing one and LFB ticket. When this happens, we should be able to force stop the finalizing block, and drain out all the queued finalizing requests.