Fix duplicate block notarization tickets verification
When receive a verify block message, we will also try to update previous block's notarization state. A round could receive N(generator num) blocks of a round, and if they all have the same previous block, then we will update the previous block' notarization state N times, which is a waste of resources.