Add StopOnBlockSyncingHandler to avoid unnecessary message process
Applied this to `/v1/_m2m/block/verification_ticket` handler, so if the miner finds it's struggling on catching up with the latest blocks, it would ignore the request for `verification_ticket`, the N2N request signature verification will also be ignored.

When the following cases are met, we will the miner is struggling on catching up the blocks.

- The current round is behind the LFB ticket
- The local LFB is 3 rounds behind the LFB ticket
- the local LFB is config.GetLFBTicketAhead() behind the current round.