Fix sendTo slow issue
The sendTo is slow because it does not return until it got the response from the remote. However, it does not care about what the response is, it just ignores all the data in response. In terms of the message receiver, it could give response slow, as it would not send response until it has processed the message, which is not necessary at all. So send response immediately after reading out the data from the request, and then verify the sender and start to process the message.