Run AddOrOverwrite for both miner and sharder
As the stake pool rewards will distribute rewards to miners and sharders from magic block, while if the network just come up, the events db does not even have the sharders added. The reward updating batch sql will insert the sharder or miner to the sharders table with most of the field empty except the rewards field, which will fail the add sharder/miner event afterwards because a new insert action will break the unique sharder id rule. Anyway, the reward distribution is kind of a bug we need to fix later.