Fix vrf cache
Set current round on StartNextRound before adding the VRF to lock the current round. Otherwise, we will miss one future VRF, and if that's the last ticket to meet threshold, we will have to wait soft round timeout.

- on round 100, receive future VRF for round 101, cache it
- start next round (still on round 100 at the moment), create VRF for round 101, and broadcast it to others
- add the created VRF to local meanwhile checking if there are cached future VRF, and process it. So now, we have 2 tickets
- received another ticket, but since the current round is still on 100, we will cache it instead of processing it.
- set current round to 101 when start next round process is finished.
- wait soft round time out to re do the VRF process

So if in StartNextRound method, we set current round to 101 before creating local VRF, the last ticket will not be treated as future ticket and we will process it immediately, and meet the threshold