Bind allocation to blobber when creating allocation
Rather than bind them on committing_connection SC, and unbind them when there's no data stored in the allocation, we connect them immediately when creating the allocation, and only unbind them when removing blobber from the allocation, or finalize/cancel the allocation. In this way, we can avoid the potential inconsistent state and avoid unnecessary add/remove operations.