Fix sharder stuck caused by missing blocks
It happens becuase current round is the fork block, and local blocks store does not have the correct notarized block that the next round block connects to. In that case, the previous process block will return instead of fetching the block. So in this fix, we will fetch it and compute its state and then re-process the block